refueling_boom: Compute receiver position in same timer as boom
The timer which computes the orientation of the boom depending on the
tracking receiver's position, got this position from a second timer.
Instead, compute the receiver position in the same first timer. All that
the second timer now does is computing which callsign is the closest.

Also make sure only one of the two timers is running at a time: either
the boom is tracking a receiver, or it is computing a candidate to be
tracked (the closest aircraft).

Signed-off-by: onox <denkpadje@gmail.com>